+++ app/src/App.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import {
   Route,
-  Switch
+  Switch,
+  Redirect
 } from 'react-router-dom'
 import withSizes from 'react-sizes'
 import { compose } from 'recompose'
@@ -10,6 +11,7 @@ import Login from './pages/Login'
 import Cashflow from './pages/Cashflow'
 import Roadmap from './pages/Roadmap'
 import Signals from './pages/Signals'
+import Detailed from './pages/Detailed'
 import Account from './pages/Account'
 import SideMenu from './components/SideMenu'
 import MobileMenu from './components/MobileMenu'
@@ -26,9 +28,11 @@ export const App = ({isDesktop}) => (
         <Route exact path='/projects' component={Cashflow} />
         <Route exact path='/roadmap' component={Roadmap} />
         <Route exact path='/signals' component={Signals} />
+        <Route exact path='/projects/:ticker' component={Detailed} />
         <Route exact path='/account' component={Account} />
-        <Route path={'/login'} component={Login} />
-        <Route exact path={'/'} component={Cashflow} />
+        <Route path='/login' component={Login} />
+        <Route exact path='/' component={Cashflow} />
+        <Redirect from='/' to='/cashflow' />
       </Switch>
     </ErrorBoundary>
   </div>